refactor(embeddedfs): move embedded package filtering to common package
Move embedded filesystem package filtering logic from annotator into extractor/filesystem/embeddedfs/common so it can be reused across components.
